+++ pyinstrument/renderers/__init__.py
@@ -2,6 +2,6 @@ from pyinstrument.renderers.base import FrameRenderer, Renderer
 from pyinstrument.renderers.console import ConsoleRenderer
 from pyinstrument.renderers.html import HTMLRenderer
 from pyinstrument.renderers.jsonrenderer import JSONRenderer
-from pyinstrument.renderers.profrenderer import ProfRenderer
+from pyinstrument.renderers.pstatrenderer import PstatRenderer
 from pyinstrument.renderers.session import SessionRenderer
 from pyinstrument.renderers.speedscope import SpeedscopeRenderer
